IR Final 1.2/college/backend/requirements.txt
# sarvamai removed — using local Ollama + LLaMA 3 instead
Flask==2.3.3
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
chromadb>=0.4.0
sentence-transformers>=2.2.0
requests>=2.28.0
httpx>=0.24.0
gunicorn>=21.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
fpdf2>=2.7.0
twilio>=8.0.0
python-dotenv>=1.0.0
